+++ app/Models/Cadastros/Produtor.php
@@ -61,5 +61,7 @@ class Produtor extends Model
     {
         return $this->belongsTo(\App\Models\Cadastros\Proprietario::class);
     }
+
+    
     
 }