+++ routes/backpack/custom.php
@@ -17,6 +17,7 @@ Route::group([
 
     // Cadastros
     Route::crud('cadastros/empresas', 'Cadastros\TenantCrudController');
+    Route::crud('cadastros/ano-agricola', 'Cadastros\AnoAgricolaCrudController');
     Route::crud('cadastros/safra', 'Cadastros\SafraCrudController');
     Route::crud('cadastros/cultura', 'Cadastros\CulturaCrudController');
     Route::crud('cadastros/variedade-cultura', 'Cadastros\VariedadeCulturaCrudController');
@@ -24,5 +25,10 @@ Route::group([
     Route::crud('cadastros/produtor', 'Cadastros\ProdutorCrudController');
     Route::crud('cadastros/fazenda', 'Cadastros\FazendaCrudController');
     Route::crud('cadastros/talhao', 'Cadastros\TalhaoCrudController');
-    Route::get('cadastros/talhao/areaTalhao/{idTalhao}', 'Cadastros\TalhaoCrudController@areaTalhao');
+    Route::get('cadastros/talhao/areaFazenda/{idFazenda}', 'Cadastros\TalhaoCrudController@areaFazenda');
+    Route::crud('cadastros/locacao-talhao', 'Cadastros\LocacaoTalhaoCrudController');
+    Route::post('cadastros/locacao-talhao/variedade', 'Cadastros\LocacaoTalhaoCrudController@variedade');
+    Route::crud('cadastros/matriz-frete', 'Cadastros\MatrizFreteCrudController');
+    Route::crud('cadastros/inscricao-estadual', 'Cadastros\InscricaoEstadualCrudController');
+    
 }); // this should be the absolute last line of this file